App/requirements.txt
mysql-connector-python
python-dotenv
pydantic
pytz
bcrypt
PyJWT
python-telegram-bot
requests